add editor layer to editor module
